requirements.txt
Flask==1.1.2
Pillow==7.2.0
requests==2.24.0
dlib==19.20.0